Update: Retry service moving it to utils and making it accept the token param
