+++ src/utils/caching/retry-token.js
@@ -1,22 +1,21 @@
-import { drop } from '../../../utils/caching/drop.js'
-import { DAL_TOKEN } from '../../../constants/cache-keys.js'
-import { UNAUTHORIZED } from '../../../constants/status-codes.js'
+import { drop } from './drop.js'
+import { UNAUTHORIZED } from '../../constants/status-codes.js'
 
 /**
  * Retries an asynchronous function multiple times if it fails
  *
- * If the function throws a 401 error, the cached DAL token is dropped so that
+ * If the function throws a 401 error, the cached token is dropped so that
  * a new token can be fetched on the next attempt
  *
  * Each time it retries, it waits a bit longer before trying again, doubling
  * the wait if exponential is true
  */
-const retry = async (fn, retriesLeft = 3, interval = 1000, exponential = true) => {
+const retry = async (fn, TOKEN, retriesLeft = 3, interval = 1000, exponential = true) => {
   try {
     return (await fn())
   } catch (err) {
     if (err.isBoom && err.output.statusCode === UNAUTHORIZED) {
-      await drop(DAL_TOKEN)
+      await drop(TOKEN)
     }
     if (retriesLeft > 0) {
       await new Promise(resolve => setTimeout(resolve, interval))
